% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ser_reg.R
\name{SER_fun}
\alias{SER_fun}
\title{Single effect Bayesian regression model}
\usage{
SER_fun(y, X, sigma, sigma_0, pi = NULL, credible_prob = 0.95)
}
\arguments{
\item{y}{vector of phenotypes standardized and centered.}

\item{X}{SNP matrix of dimenstion n by p (centered and standardized)}

\item{sigma}{prior sigma}

\item{sigma_0}{prior probablity for beta}

\item{pi}{vector of length p represent prior probability of the SNPs}

\item{credible_prob}{probability of the credible set containing atleast 1 causal SNP}
}
\description{
This function allows you to express your love of cats.
}
\examples{
SER_fun()
}
